refactor: verify code implements strict 4 tier architecture for client view
