+++ frontend/src/application/ClientManagementService.ts
@@ -1,0 +1,36 @@
+import { useState, useEffect } from "react";
+import { collection, onSnapshot, query, orderBy } from "firebase/firestore";
+import { db } from "@/firebase/firebase";
+import { type ClientDoc } from "@/components/clients/ClientList";
+
+export const useClientManagementService = () => {
+  const [allDocs, setAllDocs] = useState<ClientDoc[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<Error | null>(null);
+
+  useEffect(() => {
+    const q = query(collection(db!, "clients"), orderBy("created_at", "desc"));
+    
+    const unsubscribe = onSnapshot(
+      q,
+      (snapshot) => {
+        const docs = snapshot.docs.map((d) => ({
+          id: d.id,
+          ...d.data(),
+        })) as ClientDoc[];
+        
+        setAllDocs(docs);
+        setIsLoading(false);
+      },
+      (err) => {
+        console.error("[ClientManagementService] Core Exception:", err);
+        setError(err);
+        setIsLoading(false);
+      }
+    );
+
+    return () => unsubscribe();
+  }, []);
+
+  return { allDocs, isLoading, error };
+};
