+++ frontend/src/components/clients/ClientFilterBar.tsx
@@ -1,0 +1,52 @@
+import { IconChevronDown, IconSearch } from "@/components/ui/Icons";
+
+interface ClientFilterBarProps {
+  searchQuery: string;
+  setSearchQuery: (val: string) => void;
+  statusFilter: string;
+  setStatusFilter: (val: "all" | "registered" | "interested" | "draft") => void;
+}
+
+export default function ClientFilterBar({
+  searchQuery,
+  setSearchQuery,
+  statusFilter,
+  setStatusFilter,
+}: ClientFilterBarProps) {
+  return (
+    <div className="mb-4 flex flex-col gap-3 sm:flex-row sm:items-center">
+      {/* Search input */}
+      <div className="relative flex-1">
+        <span className="pointer-events-none absolute inset-y-0 left-3 flex items-center text-slate-400">
+          <IconSearch className="h-4 w-4" />
+        </span>
+        <input
+          type="search"
+          value={searchQuery}
+          onChange={(e) => setSearchQuery(e.target.value)}
+          placeholder="Search by name, passport ID, phone, or email…"
+          className="w-full rounded-lg border border-slate-200 bg-white py-2.5 pl-9 pr-4 text-sm text-slate-800 shadow-sm placeholder:text-slate-400 focus:border-indigo-400 focus:outline-none focus:ring-2 focus:ring-indigo-200"
+        />
+      </div>
+
+      {/* Journey-status dropdown */}
+      <div className="relative sm:w-48">
+        <select
+          value={statusFilter}
+          onChange={(e) =>
+            setStatusFilter(e.target.value as "all" | "registered" | "interested" | "draft")
+          }
+          className="w-full appearance-none rounded-lg border border-slate-200 bg-white py-2.5 pl-4 pr-9 text-sm font-medium text-slate-700 shadow-sm focus:border-indigo-400 focus:outline-none focus:ring-2 focus:ring-indigo-200"
+        >
+          <option value="all">All Statuses</option>
+          <option value="registered">Registered</option>
+          <option value="interested">Interested</option>
+          <option value="draft">Draft</option>
+        </select>
+        <span className="pointer-events-none absolute inset-y-0 right-3 flex items-center text-slate-400">
+          <IconChevronDown className="h-4 w-4" />
+        </span>
+      </div>
+    </div>
+  );
+}
